Made copyFile a helper func for copy script
